--- src/components/AddMessageCard/index.js
@@ -1,11 +1,0 @@
-import AddMessageEnabled from '../../assets/icon/AddMessageEnabled.png';
-import AddMessageFocus from '../../assets/icon/AddMessageFocus.png';
-import AddMessageHover from '../../assets/icon/AddMessageHover.png';
-import AddMessagePressed from '../../assets/icon/AddMessagePressed.png';
-
-export {
-  AddMessageEnabled,
-  AddMessageFocus,
-  AddMessageHover,
-  AddMessagePressed,
-};
