--- src/components/MessageCardModal/index.js
@@ -1,5 +1,0 @@
-import { MessageCardHeader } from '../MessageCardHeader/MessageCardHeader';
-import { PostIDContext } from '../../context/PostIDContext';
-import { getFormattedDate } from '../../assets/utils/getFormattedDate';
-
-export { MessageCardHeader, PostIDContext, getFormattedDate };
