+++ src/pages/PostIDPage/components/CardBadage/CardBadage.jsx
