+++ src/pages/PostIDPage/components/CardBadage/CardBadage.style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { COLORS } from '../../style/colorPalette';
-import { FONT_STYLE } from '../../style/fontStyle';
+import { COLORS } from '../../../../style/colorPalette';
+import { FONT_STYLE } from '../../../../style/fontStyle';
 
 const BADGE = {
   가족: { color: COLORS.GREEN_500, background: COLORS.GREEN_100 },
