+++ src/pages/PostIDPage/components/MessageCard/index.js
@@ -1,0 +1,5 @@
+import { PostIDContext } from '../../../../context/PostIDContext';
+import { getFormattedDate } from '../../../../assets/utils/getFormattedDate';
+import Deleted from '../../../../assets/icon/Deleted.png';
+
+export { PostIDContext, getFormattedDate, Deleted };
