+++ src/pages/PostIDPage/components/MessageCardModal/index.js
@@ -1,0 +1,5 @@
+import { MessageCardHeader } from '../MessageCardHeader/MessageCardHeader';
+import { PostIDContext } from '../../../../context/PostIDContext';
+import { getFormattedDate } from '../../../../assets/utils/getFormattedDate';
+
+export { MessageCardHeader, PostIDContext, getFormattedDate };
