+++ src/pages/PostIDPage/components/Toast/Toast.style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { FONT_STYLE } from '../../style/fontStyle';
-import { COLORS } from '../../style/colorPalette';
+import { FONT_STYLE } from '../../../../style/fontStyle';
+import { COLORS } from '../../../../style/colorPalette';
 
 export const ToastWrapper = styled.div`
   position: fixed;
