+++ src/pages/PostIDPage/index.js
@@ -1,7 +1,7 @@
-import { MessageCardModal } from '../../components/MessageCardModal/MessageCardModal';
-import { MessageCardWrapper } from '../../components/MessageCardWrapper/MessageCardWrapper';
-import { Scrollbar } from '../../components/Scrollbar/Scrollbar';
-import { Toast } from '../../components/Toast/Toast';
+import { MessageCardWrapper } from './components/MessageCardWrapper/MessageCardWrapper';
+import { Scrollbar } from './components/Scrollbar/Scrollbar';
+import { MessageCardModal } from './components/MessageCardModal/MessageCardModal';
+import { Toast } from './components/Toast/Toast';
 import Header from '../../components/Common/Header/Header';
 import SubHeader from '../../components/SubHeader/SubHeader';
 export {
